#459: Moved and renamed the files containing the PassiveEvasionPhysical/MagicalParser
 The Parsers were not modified.
