+++ src/app/ffbe/mappers/effects/passive-effect-parser.factory.ts
@@ -2,10 +2,10 @@ import {EffectParser} from './effect-parser';
 import {UnknownEffectParser} from './unknown-effect-parser';
 import {PassiveStatsIncreaseHpThresholdParser} from './passives/passive-stats-increase-hp-threshold.parser';
 import {PassiveMpRecoveryParser} from './passives/passive-mp-recovery.parser';
-import {PassiveEvasionPhysicalParser} from './passives/passive-evasion-physical.parser';
+import {PassiveEvasionPhysicalParser} from '../../model/effects/passives/passive-evasion-physical-effect.model';
 import {PassiveCounterAttackChanceParser} from './passives/passive-counter-attack-chance.parser';
 import {PassiveMpDecreaseForSongsParser} from './passives/passive-mp-decrease-for-songs.parser';
-import {PassiveEvasionMagicalParser} from './passives/passive-evasion-magical.parser';
+import {PassiveEvasionMagicalParser} from '../../model/effects/passives/passive-evasion-magical-effect.model';
 import {PassiveSkillMultipleActivationParser} from './passives/passive-skill-multiple-activation.parser';
 import {PassiveSkillTurnStartActivationParser} from './passives/passive-skill-turn-start-activation.parser';
 import {PassiveDeceivesDeathParser} from './passives/passive-deceives-death.parser';
